+++ CHANGELOG.md
@@ -2,13 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.1] - 2024-10-30
+
+###- fixed unicorn shutdown issue
+- added endpoint for returning count per network
 
 ## [1.2.0] - 2024-10-30
 
 ### Added
 - Added commune ai-related Validator API endpoints.
 - Add endpoint for returning available miners per network.
-
 
 ## [1.0.0] - 2024-10-23